@@ -6,7 +6,7 @@ import '../models/category.dart';
 class CategoriesMealsScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final category = ModalRoute.of(context).settings.arguments as Category;
+    final category = ModalRoute.of(context)!.settings.arguments as Category;
 
     final categoryMeals = DUMMY_MEALS
         .where((meal) => meal.categories.contains(category.id))
