@@ -32,7 +32,7 @@ class MealDetailScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final meal = ModalRoute.of(context).settings.arguments as Meal;
+    final meal = ModalRoute.of(context)!.settings.arguments as Meal;
 
     return Scaffold(
       appBar: AppBar(
